apps/pr-reviewer-mcp-host/pyproject.toml
[project]
name = "pr-assistant"
version = "0.1.0"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "httpx",
  "python-dotenv",
  "pydantic",
  "openai",
  "mcp-python",
  "loguru",
  "requests",
  "aiohttp"
]

